Rename demo package
