+++ demopackage/README.md
@@ -1,3 +1,3 @@
-# mypackage
+# demopackage
 
 Demonstration of a third party user package that can be hot reloaded.
