+++ demopackage/demopackage/app.py
@@ -1,6 +1,6 @@
 from external_package.mock_imports import external_function
 
-from mypackage.dep import local_function
+from demopackage.dep import local_function
 
 
 def run_everything():
